python-whatsapp-bot-main/requirements.txt
flask==2.3.3
openai==1.1.1
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
httpx==0.24.1
anyio==3.7.1
sniffio==1.3.0